rename util.js to utils.js
