+++ src/components/trip-info.js
@@ -1,4 +1,4 @@
-import {getShortDate} from '../util.js';
+import {getShortDate} from '../utils.js';
 
 const getDateTitle = (eventList) => {
   return `${getShortDate(eventList[0].start)}&nbsp;&mdash;&nbsp;${getShortDate(eventList[eventList.length - 1].finish)}`;
