+++ src/mock/event-data.js
@@ -1,4 +1,4 @@
-import {getRandomElement, getRandomWeek, getRandomTwoDays, getRandomNumber, getRandomBoolean} from '../util.js';
+import {getRandomElement, getRandomWeek, getRandomTwoDays, getRandomNumber, getRandomBoolean} from '../utils.js';
 import {EventType, Destinations, EVENT_DEFAULT} from '../const.js';
 import {generateOfferList} from './offer-data.js';
